==> notebooks/ebdt_data/sub_network/p_model/08-testing-exp1-marginf.ipynb <==
#### Prerequesites:
* Define schema (make figure) 
* Define structure (make figure)
* Define schema class in python (src/models/model_schemata.py)
* Convert schema + structure into model template (models/templates/p_model.pl)

#### Procedure:
* Load data
* Adjust model script for lfi accordingly
* Create evidence file (if I don't need different evidence files, move to builmodel notebook)
* LFI

Multiple instances, 1 pst, X enzymes, one sample per instance, multiple enzymes per instance, test how many enzymes can have in one instances, how computing time depends on N p::f for 60 samples

%load_ext autoreload
%autoreload 2

# Importing libraries
import pandas as pd
import numpy as np
import os
import pickle
import random

from src.models import model_schemata as schema
from src.models import build_model as build
from src.models import parameter_learning as lfi
from src.models import inference as inf
from src.visualization import visualize as viz


# os.chdir('/Users/magdalena/OneDrive - Queen Mary, University of London/bezzlab/research/projects/phospho_pi/')
os.chdir('/home/mhuebner/Desktop/bezzlab/research/projects/phospho_pi/')

#### Inference (I)

# reading csv from files into dict
data = {}
data['e_ksea'] = pd.read_csv('data/processed/ebdt_data/sub_network_p/e_ksea.csv')
data['p_fc'] = pd.read_csv('data/processed/ebdt_data/sub_network_p/p_fc.csv')

# getting sample names (union of all samples in all datasets)
samples = list(set(data['e_ksea']['sample']).union(set(data['p_fc']['sample'])))
samples.sort()
# sample x% of samples randomly without replacement with seed
random.seed(612)
train = random.sample(samples, int(len(samples)*0.8))
test = [x for x in samples if x not in train]
print(test)
# filter data
testing_data = {}
testing_data['e_ksea'] = data['e_ksea'][data['e_ksea']['sample'].isin(test)].reset_index(drop=True)
testing_data['p_fc'] = data['p_fc'][data['p_fc']['sample'].isin(test)].reset_index(drop=True)

# Adding evidence
model_inference = 'models/ebdt_data/sub_network/p_model/p_model_exp1_testing_200i.pl'

#### Inferring p_occupancy

Adding queries

sample_ids = testing_data['p_fc']['sample'].unique().tolist()
phosphosites = testing_data['p_fc']['phosphosite'].unique().tolist()
queries = inf.generate_queries(schema.POccupancyPredicate, sample_ids=sample_ids, phosphosites=phosphosites)
len(queries)

queries[:10]

Adding evidence (phosphosite and relevant enzymes)

# Mapping data to Problog predicates
predicates = {}
predicates['p_fc'] = schema.PFoldChangePredicate()
predicates['p_fc'].add_data(testing_data['p_fc'], 'phosphosite', 'sample', 'value')
predicates['e_ksea'] = schema.EKseaPredicate()
predicates['e_ksea'].add_data(testing_data['e_ksea'], 'enzyme', 'sample', 'value')

data['es_interaction'] = pd.read_csv('data/processed/ebdt_data/sub_network_p/es_interaction.csv')

evidence_dict = {}
for s in sample_ids:
    evidence_dict[s] = {}
    for p in phosphosites:
        enzymes = data['es_interaction']['enzyme'][data['es_interaction']['phosphosite'] == p].tolist()
        evid_generator = build.ProblogStatementGenerator(predicates['e_ksea'])
        evidence_e = []
        for e in enzymes:
            evidence_e.extend(evid_generator.generate_facts(build.EvidenceTemplate, select=[s, e]))
        evid_generator = build.ProblogStatementGenerator(predicates['p_fc'])
        evidence_p = evid_generator.generate_facts(build.EvidenceTemplate, select=[s, p])
        evidence_dict[s][p] = evidence_e + evidence_p

# make evidence_dict['AZD5438'] into list
evidence_list = []
for s in evidence_dict.keys():
    for p in evidence_dict[s]:
        evidence_list.append(evidence_dict[s][p])
evidence_list[:10]

inference_output = 'models/ebdt_data/sub_network/p_model/inference_log_1.txt'
inf_data_dict = inf.infer_marginf(model_inference, queries, inference_output, evidence=evidence_list)
pickle.dump(inf_data_dict, open('models/ebdt_data/sub_network/p_model/testing_relepev_exp1_200i.pkl', 'wb'))

Adding evidence (relevant enzymes only)

data['es_interaction'] = pd.read_csv('data/processed/ebdt_data/sub_network_p/es_interaction.csv')

evidence_dict = {}
for s in sample_ids:
    evidence_dict[s] = {}
    for p in phosphosites:
        enzymes = data['es_interaction']['enzyme'][data['es_interaction']['phosphosite'] == p].tolist()
        evid_generator = build.ProblogStatementGenerator(predicates['e_ksea'])
        evidence_e = []
        for e in enzymes:
            evidence_e.extend(evid_generator.generate_facts(build.EvidenceTemplate, select=[s, e]))
        evidence_dict[s][p] = evidence_e

# make evidence_dict['AZD5438'] into list
evidence_list = []
for s in evidence_dict.keys():
    for p in evidence_dict[s]:
        evidence_list.append(evidence_dict[s][p])
evidence_list[:10]

inference_output = 'models/ebdt_data/sub_network/p_model/inference_log_1.txt'
inf_data_dict = inf.infer_marginf(model_inference, queries, inference_output, evidence=evidence_list)
pickle.dump(inf_data_dict, open('models/ebdt_data/sub_network/p_model/testing_releev_exp1_200i.pkl', 'wb'))
